requirements.txt
Faker==25.2.0
numpy==1.24.4
paho-mqtt==2.1.0
Pyfhel==3.4.2
python-dateutil==2.9.0.post0
six==1.16.0
